Totally fucked up refcounting

+++ source/rt/stdc.d
@@ -3,7 +3,7 @@ module rt.stdc;
 extern (C) int printf(const char * fmt, ...);
 extern (C) int puts(char * str);
 
-extern (C) void * malloc(int);
+extern (C) void * malloc(size_t);
 extern (C) void free(void *);
 extern (C) void * memcpy(void *, const(void) *, size_t);
